requirements.txt
llama-index==0.10.57
llama-index-embeddings-huggingface==0.2.2
llama-index-vector-stores-chroma==0.1.10
chromadb==0.5.3
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyyaml==6.0.1
python-dotenv==1.0.1
pydantic==2.8.2

# MCP and development dependencies
fastapi_mcp
pre-commit
ruff
black
mypy
pytest
pytest-cov
pytest-asyncio
types-PyYAML
